rag-explorer/requirements.txt
chromadb>=0.4.22
sentence-transformers>=2.2.2
flask>=3.0.0
flask-cors>=4.0.0
PyPDF2>=3.0.0
openai>=1.12.0
requests>=2.31.0
python-dotenv>=1.0.0
